Added <Hero> class exception
Had to add this because upon importing Hero into a hero module. Inspect then asumes that <Hero> is now a class member of the new hero module. It then returns <Hero> as 1 of its members from that module.

To replicate this just make a hero inside the "heroes" folder and then check the "heroes" global in the main python file. "heroes" global now contains the <Hero> class as 1 of it items.